Yhdistä suoritusnäkymien e2e-sivuobjektit BaseSuorituksetPageen
