+++ e2e/models/BaseSuorituksetPage.ts
@@ -1,0 +1,41 @@
+import { Locator, Page } from "@playwright/test"
+import BasePage from "./BasePage"
+import { Config } from "../config"
+
+export default abstract class BaseSuorituksetPage extends BasePage {
+  constructor(page: Page, config: Config) {
+    super(page, config)
+  }
+
+  getSuorituksetTable() {
+    return this.getPageContent().getByRole("table")
+  }
+
+  abstract getSuoritusRow(): Locator
+
+  getSuoritusColumn(rowIndex: number, columnIndex: number) {
+    const row = this.getSuoritusRow().nth(rowIndex)
+    return row.getByRole("cell").nth(columnIndex)
+  }
+
+  getTableColumnHeaderLink(text: string) {
+    return this.getSuorituksetTable().getByRole("link", { name: text })
+  }
+
+  getErrorLink() {
+    return this.getPageContent().locator(".error-text").getByRole("link")
+  }
+
+  getCSVDownloadLink() {
+    return this.getPageContent().getByRole("link", {
+      name: "Lataa tiedot CSV:nä",
+    })
+  }
+
+  protected async openFilterDialogLocator(): Promise<Locator> {
+    await this.getPageContent()
+      .getByRole("button", { name: "Rajaa näytettävät tiedot" })
+      .click()
+    return this.getPageContent().getByTestId("table-filter-dialog")
+  }
+}
