+++ e2e/models/kotoutumiskoulutus/KielitestiSuorituksetPage.ts
@@ -1,10 +1,9 @@
 import { Locator, Page } from "@playwright/test"
-import BasePage from "../BasePage"
+import BaseSuorituksetPage from "../BaseSuorituksetPage"
 import { Config } from "../../config"
-import { VktSuorituksetFilterDialog } from "../vkt/VktSuorituksetFilterDialog"
 import { KielitutkintoSuorituksetFilterDialog } from "./KielitestiSuorituksetFilterDialog"
 
-export default class KielitestiSuorituksetPage extends BasePage {
+export default class KielitestiSuorituksetPage extends BaseSuorituksetPage {
   searchField: Locator
 
   constructor(page: Page, config: Config) {
@@ -24,40 +23,13 @@ export default class KielitestiSuorituksetPage extends BasePage {
     return this.page.getByRole("heading", { name: name })
   }
 
-  getSuorituksetTable() {
-    return this.getPageContent().getByRole("table")
-  }
-
-  getSuoritusRow() {
-    const suorituksetTable = this.getSuorituksetTable()
-    return suorituksetTable.getByTestId("suoritus-summary-row")
-  }
-
-  getSuoritusColumn(rowIndex: number, columnIndex: number) {
-    const row = this.getSuoritusRow().nth(rowIndex)
-    return row.getByRole("cell").nth(columnIndex)
-  }
-
-  getTableColumnHeaderLink(text: string) {
-    return this.getSuorituksetTable().getByRole("link", { name: text })
-  }
-
-  getErrorLink() {
-    return this.getPageContent().locator(".error-text").getByRole("link")
-  }
-
-  getCSVDownloadLink() {
-    return this.getPageContent().getByRole("link", {
-      name: "Lataa tiedot CSV:nä",
-    })
+  getSuoritusRow(): Locator {
+    return this.getSuorituksetTable().getByTestId("suoritus-summary-row")
   }
 
   async openFilterDialog() {
-    await this.getPageContent()
-      .getByRole("button", { name: "Rajaa näytettävät tiedot" })
-      .click()
     return new KielitutkintoSuorituksetFilterDialog(
-      this.getPageContent().getByTestId("table-filter-dialog"),
+      await this.openFilterDialogLocator(),
     )
   }
