+++ e2e/models/yki/YkiSuorituksetPage.ts
@@ -1,5 +1,5 @@
-import BasePage from "../BasePage"
-import { Page } from "@playwright/test"
+import BaseSuorituksetPage from "../BaseSuorituksetPage"
+import { Locator, Page } from "@playwright/test"
 import { expect } from "../../fixtures/baseFixture"
 import { Config } from "../../config"
 import * as node_fs from "node:fs"
@@ -7,7 +7,7 @@ import { YkiSuorituksetFilterDialog } from "./YkiSuorituksetFilterDialog"
 
 const fs = node_fs.promises
 
-export default class YkiSuorituksetPage extends BasePage {
+export default class YkiSuorituksetPage extends BaseSuorituksetPage {
   constructor(page: Page, config: Config) {
     super(page, config)
   }
@@ -28,47 +28,23 @@ export default class YkiSuorituksetPage extends BasePage {
     ).toBeVisible()
   }
 
-  getContent() {
-    return this.getPageContent()
-  }
-
-  getSuorituksetTable() {
-    return this.getContent().getByRole("table")
-  }
-
-  getSuoritusRow() {
-    const suorituksetTable = this.getSuorituksetTable()
-    return suorituksetTable.locator(".suoritus")
-  }
-
-  getErrorLink() {
-    return this.getContent().locator(".error-text").getByRole("link")
-  }
-
-  getCSVDownloadLink() {
-    return this.getPageContent().getByRole("link", {
-      name: "Lataa tiedot CSV:nä",
-    })
+  getSuoritusRow(): Locator {
+    return this.getSuorituksetTable().locator(".suoritus")
   }
 
   async openFilterDialog() {
-    await this.getContent()
-      .getByRole("button", { name: "Rajaa näytettävät tiedot" })
-      .click()
-    return new YkiSuorituksetFilterDialog(
-      this.getContent().getByTestId("table-filter-dialog"),
-    )
+    return new YkiSuorituksetFilterDialog(await this.openFilterDialogLocator())
   }
 
   async setSearchTerm(search: string) {
-    await this.getContent()
+    await this.getPageContent()
       .getByRole("search")
       .getByPlaceholder("Oppijanumero, henkilötunnus, Solki-ID tai hakusana")
       .fill(search)
   }
 
   async filterSuoritukset() {
-    await this.getContent().getByRole("button", { name: "Suodata" }).click()
+    await this.getPageContent().getByRole("button", { name: "Suodata" }).click()
   }
 
   async openSuoritusDetails(rowIndex: number = 0) {
@@ -76,15 +52,6 @@ export default class YkiSuorituksetPage extends BasePage {
       .nth(rowIndex)
       .getByRole("link", { name: "Näytä" })
       .click()
-  }
-
-  getSuoritusColumn(rowIndex: number, columnIndex: number) {
-    const row = this.getSuoritusRow().nth(rowIndex)
-    return row.getByRole("cell").nth(columnIndex)
-  }
-
-  getTableColumnHeaderLink(text: string) {
-    return this.getSuorituksetTable().getByRole("link", { name: text })
   }
 
   async downloadCSV(): Promise<string> {
